move fetch.js up one level
